API/requirements.txt
Django
djangorestframework
Pillow
opencv-python
django-cors-headers
